requirements.txt
Django>=3.2.3,<3.3.0
djangorestframework>=3.12.4,<3.13.0
psycopg2>=2.8.6,<2.9
uWSGI>=2.0.19.1,<2.1
oauth2client>=4.1.3,<4.2
gspread>=3.1.0,<3.2
freezegun>=0.3.11,<0.4
beautifulsoup4>=4.10.0,<4.11.0
django-cors-headers>=3.0.2,<3.1.0
xlrd>=1.2.0,<1.3.0
httpie>=1.0.2,<1.1.0
requests>=2.26.0,<3.0